Wallet button pulled out into it's own component
